Create FSR + Pot + Motor
This project uses an FSR to control the speed of a motor. Additionally, I added a potentiometer to control the 'sensitivity' of the FSR. Why? Imagine the FSR is a floor-mounted weight sensor meant to help a person operate a machine, like a stand mixer, with their weight. In this scenario it would be difficult for a heavy person to turn the mixer slowly or for a light person to turn the mixer quickly. The potentiometer acts as a calibration tool allowing the user to define what percentage of full power the sensor can access.